# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "albomx-comics API.",
        "title": "albomx-comics API",
        "contact": {
            "name": "Juan Espinoza",
            "email": "juanmesp@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/characters": {
            "get": {
                "description": "Get characters info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Albomxcomics"
                ],
                "summary": "get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "To who am i getting?",
                        "name": "avengerId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CharactersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/collaborators": {
            "get": {
                "description": "Get collaborators info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Albomxcomics"
                ],
                "summary": "get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "To who am i getting?",
                        "name": "avengerId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CollaboratorsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CharactersResponse": {
            "type": "object",
            "properties": {
                "characters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "lastSync": {
                    "type": "string"
                }
            }
        },
        "CollaboratorsResponse": {
            "type": "object",
            "properties": {
                "colorists": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "editors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lastSync": {
                    "type": "string"
                },
                "writers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "description": "All the available albomx-comics operations",
            "name": "Albomxcomics"
        }
    ]
}